+++ gw2-raid-stats/src/GW2RaidStats.Infrastructure/Database/RaidStatsDb.cs
@@ -32,4 +32,5 @@ public class RaidStatsDb : DataConnection
     public ITable<EventSignupEntity> EventSignups => this.GetTable<EventSignupEntity>();
     public ITable<EventReminderPreferenceEntity> EventReminderPreferences => this.GetTable<EventReminderPreferenceEntity>();
     public ITable<EventTemplateEntity> EventTemplates => this.GetTable<EventTemplateEntity>();
+    public ITable<PlayerEncounterPhaseStatEntity> PlayerEncounterPhaseStats => this.GetTable<PlayerEncounterPhaseStatEntity>();
 }